backend/requirements.txt
flask
flask-cors
numpy<2.0
matplotlib
qutip==4.7.3
scipy<1.10
google-generativeai
